api/authentication/middleware/Decode.py
from django.utils.deprecation import MiddlewareMixin
from rest_framework_simplejwt.authentication import JWTAuthentication

class GetUserID(MiddlewareMixin):

    def process_request(self, request):
        JWT_authenticator = JWTAuthentication()
        response = JWT_authenticator.authenticate(request)
        if response is not None:
            # unpacking
            user , token = response
            request.session['decoded_user'] = token.payload['user_id']